Move resources definitions to specific package
