+++ rofication/__init__.py
@@ -3,5 +3,6 @@ from ._gui import RoficationGui
 from ._handler import NotificationHandler
 from ._notification import Notification, CloseReason, Urgency
 from ._queue import NotificationQueue, NotificationQueueObserver
+from ._resource import Resource
 from ._server import NotificationServer
 from ._static import ROFICATION_UNIX_SOCK
