+++ rofication/resources/__init__.py
@@ -1,0 +1,2 @@
+from ._static import value_font, notify_some, notify_none, notify_error, \
+    value_color, label_color, critical_color, nominal_color
